addition of mbtiles support fix bugs associated with icons & ct: fix ind. incident model export to include images; add icon to groups when converting from category; fix listtree refresh; fix export dm only
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@7064 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/cybertracker/org.wcs.smart.cybertracker.patrol/src/org/wcs/smart/cybertracker/patrol/export/PatrolPackageExporter.java
@@ -195,7 +195,11 @@ public enum PatrolPackageExporter {
 		if (file != null) {
 			Path fromPath = file.getAttachmentFile().toPath();
 			String fileName = cmObject.getImageFile().getName();
+			if (cmObject.getUuid() == null) {
+				fileName = UuidUtils.uuidToString(object.getUuid());
+			}
 			Path toPath = tempDir.resolve(SharedUtils.getFilenameWithoutExtension(fileName) + "." + SharedUtils.getFilenameExtension(fromPath.getFileName().toString())); //$NON-NLS-1$
+			if (Files.exists(toPath)) return;
 			Files.copy(fromPath, toPath);
 			if (!toIncludeInZip.contains(toPath.toFile())) toIncludeInZip.add(toPath.toFile());
 		}
